requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pandas>=1.5.0
plotly>=5.18.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.35.0
torch>=2.1.0
spacy>=3.7.0
umap-learn>=0.5.5
hdbscan>=0.8.33
faker>=20.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
